[WIP] Ports: Update Zig to 0.14.0

+++ Ports/zig/package.sh
@@ -1,25 +1,36 @@
 #!/usr/bin/env -S bash ../.port_include.sh
 
 port='zig'
-version='0.12.0-dev.141+ddf5859c2'
+version='0.14.0'
 files=(
-    'https://github.com/ziglang/zig-bootstrap/archive/34644ad5032c58e39327d33d7f96d63d7c330003.tar.gz#e502ae17b01f03c627927d60b2e26b5f7f83b0e8be27b6ef55511d52e5892ccf'
-    'https://github.com/ziglang/zig/archive/ddf5859c22527c6bf5d8bb13310db996fcc58874.tar.gz#9adaf787b6233cfbe784d2d8a72398784f3742e2f5ac700cbd59ba952f9491ad'
+    "https://github.com/ziglang/zig/releases/download/${version}/zig-bootstrap-${version}.tar.xz#bf3fcb22be0b83f4791748adb567d3304779d66d7bf9b1bd557ef6c2e0232807"
+    "https://github.com/ziglang/zig/archive/refs/tags/${version}.tar.gz#b45589b782a9406a414a3cac201ca0243650f9bedcb453730c124bf6f07ab33f"
 )
 
 # The actual directory to build in.
-workdir='zig-bootstrap-34644ad5032c58e39327d33d7f96d63d7c330003'
+workdir="zig-bootstrap-${version}"
 # The newer Zig directory we move into the workdir.
-zigdir='zig-ddf5859c22527c6bf5d8bb13310db996fcc58874'
+zigdir="zig-${version}"
 
 post_fetch() {
+    # NOTE: Running this multiple times is a massive footgun as patches only get applied once, the
+    #       next time we'd end up with a clean copy of the original Zig sources.
+    if [ -f .post-fetch-executed ]; then
+        return
+    fi
+    run touch .post-fetch-executed
+
     # Move the newer version of Zig into the bootstrap
     run rm -rf zig
-    run mv "../${zigdir}" zig
+    run cp -r "../${zigdir}" zig
 
     # Copy the scripts that the build process will use
     run mkdir -p out
     run cp -r "${PORT_META_DIR}/scripts" out/
+
+    # Copy libSystem definitions which are required on macOS, once we set $ZIG_LIBC it will no
+    # longer be found in its original place
+    run cp zig/lib/libc/darwin/libSystem.tbd "${DESTDIR}/usr/lib/"
 }
 
 build() {
